+++ scripts/hunt-prey/hunted-prey-hook.js
@@ -1,5 +1,64 @@
-import { getFlag, getItemFromActor } from "../utils/utils.js";
-import { HUNTED_PREY_EFFECT_ID, HUNT_PREY_ACTION_ID } from "./hunt-prey.js";
+import { getFlag, getItem, getItemFromActor, Updates } from "../utils/utils.js";
+import { HUNTED_PREY_EFFECT_ID, HUNT_PREY_ACTION_ID, PREY_EFFECT_ID } from "./hunt-prey.js";
+
+Hooks.on(
+    "ready",
+    () => {
+        libWrapper.register(
+            "pf2e-ranged-combat",
+            "CONFIG.PF2E.Item.documentClasses.effect.prototype._onCreate",
+            function(wrapper, ...args) {
+                if (this.sourceId === HUNTED_PREY_EFFECT_ID) {
+                    const sourceActorName = this.actor.name;
+                    const targetIds = getFlag(this, "targetIds") ?? [];
+
+                    canvas.scene.tokens
+                        .filter(token => targetIds.includes(token.id))
+                        .map(token => token.actor)
+                        .filter(actor => game.user === actor.primaryUpdater)
+                        .forEach(async actor => {
+                            const preyEffectSource = await getItem(PREY_EFFECT_ID);
+                            preyEffectSource.name = `${preyEffectSource.name} (${sourceActorName})`;
+                            preyEffectSource.flags = {
+                                ...preyEffectSource.flags,
+                                "pf2e-ranged-combat": {
+                                    "sourceEffectId": this.id
+                                }
+                            };
+
+                            actor.createEmbeddedDocuments("Item", [preyEffectSource]);
+                        });
+                }
+
+                wrapper(...args);
+            },
+            "WRAPPER"
+        );
+
+        libWrapper.register(
+            "pf2e-ranged-combat",
+            "CONFIG.PF2E.Item.documentClasses.effect.prototype._onDelete",
+            function(wrapper, ...args) {
+                if (this.sourceId === HUNTED_PREY_EFFECT_ID) {
+                    const targetIds = getFlag(this, "targetIds") ?? [];
+
+                    canvas.scene.tokens
+                        .filter(token => targetIds.includes(token.id))
+                        .map(token => token.actor)
+                        .filter(actor => game.user === actor.primaryUpdater)
+                        .forEach(actor =>
+                            actor.itemTypes.effect
+                                .find(effect => effect.sourceId === PREY_EFFECT_ID && getFlag(effect, "sourceEffectId") === this.id)
+                                ?.delete()
+                        );
+                }
+
+                wrapper(...args);
+            },
+            "WRAPPER"
+        );
+    }
+);
 
 Hooks.on(
     "targetToken",